css files moved into the styles folder
