+++ src/Styles/Home.css
@@ -1,0 +1,61 @@
+.page-container .top-section {
+    padding: 2rem 20rem;
+}
+
+.top-section {
+    display: grid;
+    grid-template-columns: 1fr 3fr;
+    gap: 25px;
+}
+
+.genres {
+    grid-column: span 1;
+}
+
+.trailer-area {
+    grid-column: span 1;
+}
+
+
+@media (max-width: 1600px) {
+    .page-container {
+        padding: 2rem 20rem;
+    }
+
+}
+
+@media (max-width: 1100px) {
+    .page-container {
+        padding: 2rem 5rem;
+    }
+
+    .movie-wrapper {
+        grid-template-columns: repeat(3, 1fr);
+    }
+
+}
+
+@media (max-width: 800px) {
+    .page-container {
+        padding: 2rem 2rem;
+    }
+
+    .movie-wrapper {
+        grid-template-columns: repeat(2, 1fr);
+    }
+
+    .top-section {
+        grid-template-columns: 1fr;
+    }
+
+    .genres,
+    .trailer-area {
+        grid-column: span 2;
+    }
+}
+
+@media (max-width: 600px) {
+    .page-container {
+        padding: 1rem;
+    }
+}
