+++ src/index.css
@@ -1,8 +1,5 @@
 
 :root {
 
-}
-
-* {
     font-family:'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif
 }
